requirements.txt
pandas==2.0.3
google==3.0.0
googlesearch-python==1.2.3
djangorestframework==3.14.0
Django==4.2.6
django-rest-swagger==2.2.0
drf-yasg==1.21.7
beautifulsoup4==4.12.2
cchardet==2.1.7
certifi==2023.11.17
charset-normalizer==3.3.2
cssselect==1.2.0
filelock==3.13.1
free-proxy==1.1.1
idna==3.4
lxml==4.9.3
pyquery==2.0.0
PySocks==1.7.1
requests==2.31.0
requests-file==1.5.1
six==1.16.0
soupsieve==2.5
tldextract==5.1.1
urllib3==2.1.0
celery==5.3.5
redis==5.0.1
python-dotenv==1.0.0
